refacor: move logic to ethereum
