+++ src/ape/api/networks.py
@@ -611,6 +611,7 @@ class EcosystemAPI(ExtraAttributesMixin, BaseInterfaceModel):
             Union[Type, Sequence]: The Python types for the given ABI type.
         """
 
+    @raises_not_implemented
     def decode_custom_error(
         self,
         data: HexBytes,
@@ -630,45 +631,6 @@ class EcosystemAPI(ExtraAttributesMixin, BaseInterfaceModel):
         Returns:
             Optional[CustomError]: If it able to decode one, else ``None``.
         """
-
-        # Use an instance (required for proper error caching).
-        contract = self.chain_manager.contracts.instance_at(address)
-
-        selector = data[:4]
-        input_data = data[4:]
-
-        abi = None
-        if selector not in contract.contract_type.errors:
-            # ABI not found. Try looking at the "last" contract.
-            if not (tx := kwargs.get("txn")) or not self.network_manager.active_provider:
-                return None
-            elif not (last_addr := self._get_last_address_from_trace(tx.txn_hash)):
-                return None
-            elif not (cerr := self.decode_custom_error(data, last_addr)):
-                return cerr
-
-            # error never found.
-            return None
-
-        abi = contract.contract_type.errors[selector]
-        error_cls = contract.get_error_by_signature(abi.signature)
-        inputs = self.decode_calldata(abi, input_data)
-        kwargs["contract_address"] = address
-        return error_cls(abi, inputs, **kwargs)
-
-    def _get_last_address_from_trace(self, txn_hash: Union[str, HexBytes]) -> Optional[AddressType]:
-        try:
-            trace = list(self.chain_manager.provider.get_transaction_trace(txn_hash))
-        except Exception:
-            return None
-
-        for frame in trace[::-1]:
-            if not (addr := frame.contract_address):
-                continue
-
-            return addr
-
-        return None
 
 
 class ProviderContextManager(ManagerAccessMixin):
